This crate will contain only ecc functionality.
