--- src/suites/mod.rs
@@ -1,20 +1,0 @@
-pub mod mont;
-pub mod ted;
-pub mod weier;
-
-pub use crate::suites::weier::{
-    P256_SHA256_SSWU_NU_, P256_SHA256_SSWU_RO_, P256_SHA256_SVDW_NU_, P256_SHA256_SVDW_RO_,
-    P384_SHA512_SSWU_NU_, P384_SHA512_SSWU_RO_, P384_SHA512_SVDW_NU_, P384_SHA512_SVDW_RO_,
-    P521_SHA512_SSWU_NU_, P521_SHA512_SSWU_RO_, P521_SHA512_SVDW_NU_, P521_SHA512_SVDW_RO_,
-    SECP256K1_SHA256_SSWU_NU_, SECP256K1_SHA256_SSWU_RO_,
-};
-
-pub use crate::suites::mont::{
-    CURVE25519_SHA256_ELL2_NU_, CURVE25519_SHA256_ELL2_RO_, CURVE25519_SHA512_ELL2_NU_,
-    CURVE25519_SHA512_ELL2_RO_, CURVE448_SHA512_ELL2_NU_, CURVE448_SHA512_ELL2_RO_,
-};
-
-pub use crate::suites::ted::{
-    EDWARDS25519_SHA256_EDELL2_NU_, EDWARDS25519_SHA256_EDELL2_RO_, EDWARDS25519_SHA512_EDELL2_NU_,
-    EDWARDS25519_SHA512_EDELL2_RO_, EDWARDS448_SHA512_EDELL2_NU_, EDWARDS448_SHA512_EDELL2_RO_,
-};
